refactor: StrategyRuntimeState -> StrategyInfo
