+++ src/lib/trade-executor/components/StrategyIcon.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-	import type { StrategyRuntimeState } from 'trade-executor/strategy/runtime-state';
+	import type { StrategyRuntimeState } from 'trade-executor/models/strategy-info';
 
 	export let strategy: StrategyRuntimeState;
