+++ src/routes/+page.server.ts
@@ -1,4 +1,4 @@
-import { getCachedStrategies } from 'trade-executor/strategy/runtime-state';
+import { getCachedStrategies } from 'trade-executor/models/strategy-info';
 
 export async function load({ fetch }) {
 	const strategies = (await getCachedStrategies(fetch)).filter((s) => s.frontpage);
