+++ src/routes/strategies/StrategyTile.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
 	import type { EventHandler } from 'svelte/elements';
-	import type { StrategyRuntimeState } from 'trade-executor/strategy/runtime-state';
+	import type { StrategyRuntimeState } from 'trade-executor/models/strategy-info';
 	import { goto } from '$app/navigation';
 	import { utcDay } from 'd3-time';
 	import { normalizeDataForInterval } from '$lib/chart';
